Move enumerator to partial classes
