+++ src/Lumina/Excel/Collection.cs
@@ -10,7 +10,7 @@ namespace Lumina.Excel;
 /// </summary>
 /// <remarks>Mostly an implementation detail for reading Excel rows. This type does not store or hold any row data, and is therefore lightweight and trivially constructable.</remarks>
 /// <typeparam name="T">A type that wraps a group of fields inside a row.</typeparam>
-public readonly struct Collection< T >( ExcelPage page, uint parentOffset, uint offset, Func< ExcelPage, uint, uint, uint, T > ctor, int size )
+public readonly partial struct Collection< T >( ExcelPage page, uint parentOffset, uint offset, Func< ExcelPage, uint, uint, uint, T > ctor, int size )
     : IList< T >, IReadOnlyList< T > where T : struct
 {
     /// <inheritdoc cref="ICollection{T}.Count"/>
@@ -83,39 +83,4 @@ public readonly struct Collection< T >( ExcelPage page, uint parentOffset, uint 
     /// <param name="index">Index of the item.</param>
     /// <returns>Newly created item.</returns>
     private T UnsafeCreateAt( int index ) => ctor( page, parentOffset, offset, unchecked( (uint) index ) );
-
-    /// <summary>Enumerator that enumerates over the different items.</summary>
-    /// <param name="collection">Collection to iterate over.</param>
-    public struct Enumerator( Collection< T > collection ) : IEnumerator< T >
-    {
-        private int _index = -1;
-
-        /// <inheritdoc cref="IEnumerator{T}.Current"/>
-        public T Current { get; private set; }
-
-        readonly object IEnumerator.Current => Current;
-
-        /// <inheritdoc/>
-        public bool MoveNext()
-        {
-            if( ++_index < collection.Count )
-            {
-                // UnsafeCreateAt must be called only when the preconditions are validated.
-                // If it is to be called on-demand from get_Current, then it may end up being called with invalid parameters,
-                // so we create the instance in advance here.
-                Current = collection.UnsafeCreateAt( _index );
-                return true;
-            }
-
-            --_index;
-            return false;
-        }
-
-        /// <inheritdoc/>
-        public void Reset() => _index = -1;
-
-        /// <inheritdoc/>
-        public readonly void Dispose()
-        { }
-    }
 }
